# packages/contracts/openapi/labelhub.yaml
openapi: 3.0.3
info:
  title: LabelHub API
  version: 0.10.0
  description: Contract-first API for the LabelHub modular monolith and AI worker.
servers:
  - url: http://localhost:8080/api
...
security:
  - bearerAuth: []
paths:
...
  /ai-review/rules:
    get:
      tags: [AIReview]
      operationId: listAiReviewRules
      parameters:
        - in: query
          name: taskId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: AI review rules for the task, ordered by version number.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AiReviewRule'
        '400':
          $ref: '#/components/responses/ErrorBadRequest'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
    post:
      tags: [AIReview]
      operationId: saveAiReviewRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiReviewRuleRequest'
      responses:
        '201':
          description: AI review rule saved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiReviewRule'
        '400':
          $ref: '#/components/responses/ErrorBadRequest'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
  /ai-review/rules/{ruleId}/publish:
    post:
      tags: [AIReview]
      operationId: publishAiReviewRule
      parameters:
        - in: path
          name: ruleId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: AI review rule published and activated for the task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiReviewRule'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '404':
          description: Requested resource does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /ai-review/field-assist:
    post:
      tags: [AIReview]
      operationId: createFieldAssistCall
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FieldAssistRequest'
      responses:
        '202':
          description: Field-level AI assistance accepted and traceable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FieldAssistResponse'
  /submissions/{submissionId}/ai-review:
    post:
      tags: [AIReview]
      operationId: triggerSubmissionAiReview
      parameters:
        - $ref: '#/components/parameters/SubmissionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TriggerAiReviewRequest'
      responses:
        '201':
          description: AI review completed or reused through idempotency.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiReviewResult'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
        '409':
          $ref: '#/components/responses/ErrorStateConflict'
        '502':
          description: AI provider failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    get:
      tags: [AIReview]
      operationId: getSubmissionAiProvenance
      parameters:
        - $ref: '#/components/parameters/SubmissionId'
      responses:
        '200':
          description: AI provenance for a submission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionAiProvenance'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
...
  /submissions/{submissionId}/ai-prereview/enqueue:
    post:
      tags: [AIReview]
      operationId: enqueueSubmissionAiPrereview
      parameters:
        - $ref: '#/components/parameters/SubmissionId'
      responses:
        '202':
          description: The submission was accepted into the asynchronous AI prereview queue when eligible.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskAiPrereviewEnqueueResult'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
  /submissions/{submissionId}/ai-trace:
    get:
      tags: [AIReview]
      operationId: getSubmissionAiTrace
      parameters:
        - $ref: '#/components/parameters/SubmissionId'
      responses:
        '200':
          description: Full AI provenance chain for a submission, including field-level assistance history.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionAiProvenance'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    SubmissionId:
      name: submissionId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
    ErrorBadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ErrorUnauthorized:
      description: Missing, expired, or invalid authentication.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ErrorForbidden:
      description: Authenticated user is not allowed to perform this action.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ErrorNotFound:
      description: Requested resource does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ErrorStateConflict:
      description: Request conflicts with the current business state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/ApiFieldError'
    ApiFieldError:
      description: |
        Field-level error from input validation or business guards.
        `field` uses the OpenAPI camelCase property name, for example `quotaTotal` or `deadlineAt`,
        and must match the request body schema field name.
      type: object
      required: [field, message]
      properties:
        field:
          description: OpenAPI camelCase property name that failed validation.
          type: string
        message:
          description: Human-readable field-level validation or business-guard message.
          type: string
...
    TaskAiPrereviewSummary:
      type: object
      required:
        - taskId
        - totalCount
        - pendingCount
        - processingCount
        - completedCount
        - failedCount
        - enqueueableCount
      properties:
        taskId:
          type: integer
          format: int64
        totalCount:
          type: integer
          format: int64
        pendingCount:
          type: integer
          format: int64
        processingCount:
          type: integer
          format: int64
        completedCount:
          type: integer
          format: int64
        failedCount:
          type: integer
          format: int64
        enqueueableCount:
          type: integer
          format: int64
    TaskAiPrereviewEnqueueResult:
      type: object
      required: [taskId, enqueuedCount, skippedCount, summary]
      properties:
        taskId:
          type: integer
          format: int64
        enqueuedCount:
          type: integer
          format: int64
        skippedCount:
          type: integer
          format: int64
        summary:
          $ref: '#/components/schemas/TaskAiPrereviewSummary'
...
    AiReviewRuleStatus:
      type: string
      enum: [draft, published]
    AiReviewRuleRequest:
      type: object
      required: [taskId, promptTemplate, dimensions, passThreshold, rejectThreshold]
      properties:
        taskId:
          type: integer
          format: int64
        promptTemplate:
          type: string
        dimensions:
          type: array
          items:
            type: string
        threshold:
          type: number
          nullable: true
          deprecated: true
          description: Legacy pass threshold alias retained for v1 clients. New clients should send passThreshold.
        passThreshold:
          type: number
          minimum: 0
          maximum: 1
        rejectThreshold:
          type: number
          minimum: 0
          maximum: 1
    AiReviewRule:
      type: object
      required: [id, taskId, versionNo, promptVersionId, promptTemplate, dimensions, threshold, passThreshold, rejectThreshold, status, isCurrent, createdAt]
      properties:
        id:
          type: integer
          format: int64
        taskId:
          type: integer
          format: int64
        versionNo:
          type: integer
        promptVersionId:
          type: integer
          format: int64
        promptTemplate:
          type: string
        dimensions:
          type: array
          items:
            type: string
        threshold:
          type: number
          description: Legacy pass threshold alias retained for v1 clients. Mirrors passThreshold when present.
        passThreshold:
          type: number
          minimum: 0
          maximum: 1
        rejectThreshold:
          type: number
          minimum: 0
          maximum: 1
        status:
          $ref: '#/components/schemas/AiReviewRuleStatus'
        isCurrent:
          type: boolean
          description: Server-derived. True when this rule id equals the task's current_ai_review_rule_id.
        createdAt:
          type: string
          format: date-time
        activatedAt:
          type: string
          format: date-time
          nullable: true
...
    AiCallStatus:
      type: string
      enum: [completed, failed]
    AiCall:
      type: object
      required: [id, submissionId, purpose, promptVersion, providerAdapterVersion, providerName, modelName, inputHash, status, idempotencyKey, createdAt]
      properties:
        id:
          type: integer
          format: int64
        submissionId:
          type: integer
          format: int64
        fieldPath:
          type: string
          nullable: true
        purpose:
          type: string
        promptVersion:
          type: string
        promptVersionId:
          type: integer
          format: int64
          nullable: true
        aiReviewRuleId:
          type: integer
          format: int64
          nullable: true
        providerAdapterVersion:
          type: string
        providerName:
          type: string
        modelName:
          type: string
        inputHash:
          type: string
          minLength: 64
          maxLength: 64
        outputHash:
          type: string
          minLength: 64
          maxLength: 64
          nullable: true
        tokenInput:
          type: integer
        tokenOutput:
          type: integer
        cost:
          type: string
          format: decimal
        latencyMs:
          type: integer
          nullable: true
        status:
          $ref: '#/components/schemas/AiCallStatus'
        idempotencyKey:
          type: string
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        businessPrompt:
          type: string
          nullable: true
        renderedPrompt:
          type: string
          nullable: true
        requestPayload:
          type: object
          nullable: true
          additionalProperties: true
        responsePayload:
          type: object
          nullable: true
          additionalProperties: true
    AiCallInField:
      type: object
      required: [id, submissionId, fieldPath, aiCallId, accepted, userModifiedAfter, ordinal, createdAt]
      properties:
        id:
          type: integer
          format: int64
        submissionId:
          type: integer
          format: int64
        fieldPath:
          type: string
          description: StableId path, e.g. parentStableId.childStableId.
        aiCallId:
          type: integer
          format: int64
        accepted:
          type: boolean
        userModifiedAfter:
          type: boolean
        ordinal:
          type: integer
        createdAt:
          type: string
          format: date-time
    FieldFinding:
      type: object
      required: [fieldPath, severity, finding]
      properties:
        fieldPath:
          type: string
        stableId:
          type: string
          nullable: true
        label:
          type: string
          nullable: true
        severity:
          type: string
          enum: [info, warning, error]
        finding:
          type: string
        confidence:
          type: string
          format: decimal
          nullable: true
    AiCallUsage:
      type: object
      description: Provider-reported token usage. Null fields mean the provider omitted that metric.
      properties:
        promptTokens:
          type: integer
          nullable: true
        completionTokens:
          type: integer
          nullable: true
        totalTokens:
          type: integer
          nullable: true
        cacheHitTokens:
          type: integer
          nullable: true
    AiReviewResult:
      type: object
      required: [aiCall, fieldFindings, overallSuggestion, idempotencyHit]
      properties:
        aiCall:
          $ref: '#/components/schemas/AiCall'
        fieldFindings:
          type: array
          items:
            $ref: '#/components/schemas/FieldFinding'
        overallSuggestion:
          type: string
          enum: [pass, reject, manual_review]
        confidence:
          type: string
          format: decimal
          nullable: true
        summary:
          type: string
          nullable: true
        dimensionScores:
          type: array
          items:
            $ref: '#/components/schemas/DimensionScore'
        idempotencyHit:
          type: boolean
        usage:
          $ref: '#/components/schemas/AiCallUsage'
    DimensionScore:
      type: object
      required: [dimension, score]
      properties:
        dimension:
          type: string
        score:
          type: string
          format: decimal
        reason:
          type: string
          nullable: true
    TriggerAiReviewRequest:
      type: object
      required: [promptVersionId]
      properties:
        promptVersionId:
          type: integer
          format: int64
    FieldAssistRequest:
      type: object
      required: [sessionId, fieldPath, input]
      properties:
        sessionId:
          type: integer
          format: int64
        fieldPath:
          type: string
        input:
          type: object
          additionalProperties: true
        promptVersion:
          type: string
    FieldAssistResponse:
      type: object
      required: [aiCallId, fieldPath, output, provenance]
      properties:
        aiCallId:
          type: integer
          format: int64
        fieldPath:
          type: string
        output:
          type: object
          additionalProperties: true
        provenance:
          $ref: '#/components/schemas/FieldAITrace'
...
    FieldAITrace:
      type: object
      required: [aiCallId, fieldPath, source, accepted, userModifiedAfter, ordinal]
      properties:
        aiCallId:
          type: integer
          format: int64
        fieldPath:
          type: string
        source:
          type: string
          description: Prompt or provider source, such as prompt_v3.
        confidence:
          type: number
        accepted:
          type: boolean
        userModifiedAfter:
          type: boolean
        ordinal:
          type: integer
    SubmissionAiProvenance:
      type: object
      required: [submissionId, aiCalls, fieldFindings]
      properties:
        submissionId:
          type: integer
          format: int64
        aiCalls:
          type: array
          items:
            $ref: '#/components/schemas/AiCall'
        fieldFindings:
          type: array
          items:
            $ref: '#/components/schemas/AiCallInField'
...